# Anoryx-AI-Orchestrator/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Anoryx Orchestrator Internal API
  version: "1.0.0"
  description: |
    Internal, product-to-product API for Anoryx-AI-Orchestrator — the central
    orchestration layer between Anoryx Sentinel (the zero-trust AI gateway, which
    emits governance/usage events UP) and Delta (FinOps/budget policy, which submits
    policies DOWN for distribution to Sentinels).

    SCOPE OF THIS DOCUMENT (O-001 + O-002): O-001 defined THREE seams plus the
    transport security scheme; O-002 adds the cross-product event ENVELOPE that wraps the
    ingest payload, plus bounded replay, dead-letter (DLQ), and schema-version negotiation
    (see the O-002 section below). It is CONTRACT-ONLY — there is no runtime, persistence,
    or enforcement code behind it yet. The ingest pipeline, distribution engine, registry,
    persistence, UI, and deploy layers are later tasks (O-003…O-008) and MUST NOT be
    inferred from this document.

    ## The three seams
      1. Event ingest (Sentinel → Orchestrator): Sentinel pushes events up via an
         HMAC-signed webhook POST. Bodies conform to the O-002 event envelope
         (event-envelope.schema.json), whose payload wraps the LOCKED Sentinel event
         schema (contracts/events.schema.json, $id `sentinel:events:v1`) UNMODIFIED.
      2. Policy distribution intake (Delta → Orchestrator): Delta submits a policy
         record (conforming to the LOCKED contracts/policy.schema.json, $id
...
         surface.
...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - mutualTLS: []
...
paths:
...
  /v1/policies/distributions/{distribution_id}:
    get:
      operationId: getPolicyDistributionStatus
      tags: [query]
      summary: Read the distribution status of a submitted policy (read-only metadata)
      description: |
        Returns the distribution status for a previously submitted policy. READ-ONLY
        METADATA (honesty boundary c): per-target delivery state, no policy body, no
        mutation.

        NOTE: this GET-status read stays COARSE relay auth (O-006, honesty boundary d);
        per-tenant read authorization is DEFERRED (O-004 LOW-1 carried forward). It is gated by
        the shared `ORCH_SERVICE_TOKEN`; the live Delta budget-engine consumer reads this seam
        coarsely as a trusted multi-tenant relay, so a per-tenant read gate would 401 the relay.
        Only the NEW read seams (/v1/events, /v1/bus/dlq) are tenant-scoped in O-006.

        Auth: mTLS (peer = Delta) AND the coarse `ORCH_SERVICE_TOKEN` (bearer).
      security:
        - mutualTLS: []
          serviceToken: []
      parameters:
        - $ref: '#/components/parameters/DistributionId'
      responses:
        '200':
          description: The distribution status record (metadata only).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionStatus'
              examples:
                partial:
                  value:
                    distribution_id: "a7b8c9d0-1234-4ef0-9abc-def012345678"
                    policy_id: "e5f6a7b8-0123-4cde-9f01-23456789abcd"
                    state: partial
                    created_at: "2026-06-26T12:00:05Z"
                    targets:
                      - sentinel_id: "sentinel-prod-eu-1"
                        state: distributed
                        last_attempt_at: "2026-06-26T12:00:06Z"
                      - sentinel_id: "sentinel-prod-us-1"
                        state: failed
                        last_attempt_at: "2026-06-26T12:00:06Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/events:
    get:
      operationId: queryEvents
      tags: [query]
      summary: Query ingested event metadata (read-only metadata)
      description: |
        Returns a metadata-only, cursor-paginated list of ingested events for Delta.
        READ-ONLY METADATA (honesty boundary c): join keys + type + time only, never
        full event payloads. Filtering by tenant/team/agent/type/time is supported.

        NOTE: per-tenant read authorization is now ENFORCED (O-006, honesty boundary d).
        The caller's per-tenant service token establishes the principal; the read runs under
        that tenant's RLS session, so results are the principal's tenant only. A `tenant_id`
        filter that differs from the principal is rejected (403) — it is now an authorization
        boundary, not merely a convenience.

        Auth: mTLS (peer = Delta) AND a per-tenant service token (bearer).
      security:
        - mutualTLS: []
          serviceToken: []
      parameters:
        - $ref: '#/components/parameters/FilterTenantId'
        - $ref: '#/components/parameters/FilterTeamId'
        - $ref: '#/components/parameters/FilterAgentId'
        - $ref: '#/components/parameters/FilterEventType'
        - $ref: '#/components/parameters/FilterSince'
        - $ref: '#/components/parameters/FilterUntil'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of event metadata records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMetadataPage'
              examples:
                page:
                  value:
                    data:
                      - event_id: "c1d2e3f4-5678-4abc-9def-0123456789ab"
                        event_type: policy_decision_deny
                        event_timestamp: "2026-06-26T12:00:00Z"
                        tenant_id: "2a4f8c1e-0012-4b3d-9abc-d1e2f3a4b5c6"
                        team_id: "7d9e2f3a-1234-5c6b-8def-0123456789ab"
                        project_id: "b3c4d5e6-abcd-1234-ef01-234567890abc"
                        agent_id: "gateway-core"
                        request_id: "req-o001-ingest-example-01"
                    next_cursor: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: A malformed or over-length (>512 char) pagination cursor.
...
  /v1/admin/events/recent:
    get:
      operationId: listRecentEventsAdmin
      tags: [admin]
      summary: List recently ingested events across ALL tenants (operator-scoped)
      description: |
        Returns a metadata-only view of the most recently ingested events across ALL
        tenants for operator fleet triage. This is CROSS-TENANT BY DESIGN — the deliberate
        INVERSE of the O-006 PER-TENANT `/v1/events` seam: `/v1/events` derives a single
        tenant principal from a per-tenant service token and reads under that tenant's RLS
        session; this seam is gated by the OPERATOR bearer token and reads across every
        tenant so an operator can see the whole fleet at once.

        METADATA-ONLY: it reuses the SAME EventMetadata projection (join keys + type + time)
        as `/v1/events` and NEVER returns a full event payload (mirrors query honesty
        boundary c).

        NO CURSOR PAGINATION: this is a single bounded "recent N" page, newest-first,
        capped by `limit`. The absence of a cursor is a DELIBERATE SIMPLIFICATION for a
        triage surface, not an omission — an operator scanning recent activity does not
        page a full history here; deep, filtered, paginated reads remain the per-tenant
        `/v1/events` seam.

        Auth: mTLS (operator host) AND an operator bearer token.
      security:
        - mutualTLS: []
          operatorBearer: []
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A single bounded page of recent event metadata across all tenants (newest-first).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventMetadataPage'
              examples:
                recent:
                  value:
                    data:
                      - event_id: "c1d2e3f4-5678-4abc-9def-0123456789ab"
                        event_type: policy_decision_deny
                        event_timestamp: "2026-06-26T12:00:00Z"
                        tenant_id: "2a4f8c1e-0012-4b3d-9abc-d1e2f3a4b5c6"
                        team_id: "7d9e2f3a-1234-5c6b-8def-0123456789ab"
                        project_id: "b3c4d5e6-abcd-1234-ef01-234567890abc"
                        agent_id: "gateway-core"
                        request_id: "req-o001-ingest-example-01"
                      - event_id: "d2e3f4a5-6789-4bcd-8ef0-123456789abc"
                        event_type: usage_recorded
                        event_timestamp: "2026-06-26T11:59:58Z"
                        tenant_id: "9f8e7d6c-5b4a-4039-8281-706f5e4d3c2b"
                        team_id: "1a2b3c4d-5e6f-4708-8192-a3b4c5d6e7f8"
                        project_id: "0f1e2d3c-4b5a-4697-8586-7574635241f0"
                        agent_id: "gateway-core"
                        request_id: "req-o007-admin-example-02"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/distributions/recent:
    get:
      operationId: listRecentDistributionsAdmin
      tags: [admin]
      summary: List recent policy distributions across ALL tenants (operator-scoped)
      description: |
        Returns a cross-tenant SUMMARY view of the most recent policy distributions across
        ALL tenants for operator fleet triage. CROSS-TENANT BY DESIGN and gated by the
        OPERATOR bearer token, the fleet-wide counterpart to the per-distribution
        `GET /v1/policies/distributions/{distribution_id}` status seam.

        This view NEVER returns the signed policy body — no `signed_record`, no
        `content_hash` — mirroring the "no payload" discipline of the event reads. It
        exposes only join-key + classification + lifecycle-state summary fields.

        NO CURSOR PAGINATION: a single bounded "recent N" page, newest-first, capped by
        `limit` — a deliberate simplification for a triage surface, not an omission. To see
        one distribution's FULL per-target delivery status, an operator drills into the
        EXISTING `GET /v1/policies/distributions/{distribution_id}` seam, which is UNCHANGED
        and still gated by the peer service token (the coarse `ORCH_SERVICE_TOKEN`), NOT
        this operator token.

        Auth: mTLS (operator host) AND an operator bearer token.
      security:
        - mutualTLS: []
          operatorBearer: []
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A single bounded page of recent distribution summaries across all tenants (newest-first).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDistributionSummaryPage'
              examples:
                recent:
                  value:
                    data:
                      - distribution_id: "a7b8c9d0-1234-4ef0-9abc-def012345678"
                        policy_id: "e5f6a7b8-0123-4cde-9f01-23456789abcd"
                        tenant_id: "2a4f8c1e-0012-4b3d-9abc-d1e2f3a4b5c6"
                        policy_type: budget_limit
                        state: partial
                        created_at: "2026-06-26T12:00:05Z"
                      - distribution_id: "b8c9d0e1-2345-4f01-8abc-ef0123456789"
                        policy_id: "f6a7b8c9-1234-4def-8012-3456789abcde"
                        tenant_id: "9f8e7d6c-5b4a-4039-8281-706f5e4d3c2b"
                        policy_type: model_approval
                        state: distributed
                        created_at: "2026-06-26T11:58:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/identity/events:
    post:
      operationId: ingestIdentityEvent
      tags: [identity]
      summary: Ingest one normalized cross-product identity/access event
      description: |
        Sentinel, Delta, or Rendly each push ONE normalized "who accessed what, where"
        record after their OWN existing auth boundary fires — Sentinel's F-014 SSO login,
        Delta's admin-token use, Rendly's JWT verification. `source_product` is
        SERVER-RESOLVED from the matched identity-source bearer — never accepted from the
        body.

        IDEMPOTENT: a retried push with the same (source_product, idempotency_key) is
        recorded as `disposition: duplicate` (no second row, never an error). Every ingest
        ATTEMPT — fresh accept or duplicate — is hash-chain audited.

        HONESTY BOUNDARY (ADR-0010): this seam does NOT verify, translate, or federate any
        product's credential — it only durably records that a (product-verified) access
        happened. `principal_id` is an OPAQUE identifier in the source product's OWN
        namespace (an admin_user_id, an idp_subject, a static admin slug); no cross-product
        principal-identity mapping is implied or performed.

        Auth: mTLS (peer = the source product) AND an identity-source bearer.
      security:
        - mutualTLS: []
          identitySourceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityEventIngestRequest'
            examples:
              ssoLogin:
                summary: Sentinel records an operator SSO login
                value:
                  tenant_id: "2a4f8c1e-0012-4b3d-9abc-d1e2f3a4b5c6"
                  principal_type: operator
                  principal_id: "idp-subject-a1b2c3"
                  action: sso_login
                  target: "admin-console"
                  idempotency_key: "sentinel-sso-a1b2c3-1781784000"
                  occurred_at: "2026-07-08T12:00:00Z"
      responses:
        '202':
          description: The event was durably recorded (fresh accept or idempotent duplicate).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityEventIngestAccepted'
              examples:
                accepted:
                  value:
                    status: accepted
                    disposition: accepted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Request body exceeds the maximum allowed size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    get:
      operationId: queryIdentityEvents
      tags: [identity]
      summary: Query this tenant's cross-product identity events (read-only metadata)
      description: |
        Cursor-paginated read of the caller's OWN tenant's identity/access events across
        Sentinel, Delta, and Rendly. Reuses the EXISTING O-006 per-tenant principal
        (`query_service_tokens`) — the SAME credential that already gates `GET /v1/events` —
        rather than a new tenant-read trust root. Results are RLS-scoped to the principal's
        tenant; there is no cross-tenant `tenant_id` filter (unlike `/v1/events`) because
        the principal already fixes the tenant.

        Auth: mTLS (peer = Delta) AND a per-tenant service token (bearer).
      security:
        - mutualTLS: []
          serviceToken: []
      parameters:
        - name: source_product
          in: query
          required: false
          description: Filter to one source product.
          schema:
            type: string
            enum: [sentinel, delta, rendly]
        - name: principal_type
          in: query
          required: false
          schema:
            type: string
            enum: [operator, tenant_user, service_account, peer_credential]
        - name: action
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A cursor-paginated page of this tenant's identity events (ascending sequence order).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityEventPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/admin/identity/events/recent:
    get:
      operationId: listRecentIdentityEventsAdmin
      tags: [admin, identity]
      summary: List recent cross-product identity events across ALL tenants (operator-scoped)
      description: |
        Cross-tenant operator fleet-triage view of recent identity/access events — the same
        shape and the SAME operator bearer (`ORCH_ADMIN_TOKEN`) as the O-007 admin reads. A
        tenant's own scoped view is `GET /v1/identity/events`.

        Auth: mTLS (operator host) AND an operator bearer token.
      security:
        - mutualTLS: []
          operatorBearer: []
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A single bounded page of recent identity events across all tenants (newest-first).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminIdentityEventPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [admin]
      summary: A bounded operator command-dashboard summary (O-014)
      description: |
        Aggregates data the Orchestrator ALREADY tracks via its own existing seams — no new
        outbound calls, no reach into another product's tree: registry health/enabled counts
        (O-005), a distribution-state breakdown over the `limit` most-recent policy
        distributions (O-004, the same bounded page `/v1/admin/distributions/recent` already
        serves), and the `limit` most-recent `auto_disable` registry-audit circuit-breaker
        trips (O-014).

        HONESTY BOUNDARY (ADR-0014): this is NOT "system health, API loads, and governance
        metrics across all [Anoryx] products" — the Orchestrator has no access to Sentinel's,
        Delta's, or Rendly's internals beyond what already flows through its own ingest/
        registry/distribution seams. It is a real, honest summary of what the Orchestrator
        itself already knows, not a cross-product command center.

        Auth: mTLS (operator host) AND an operator bearer token.
      security:
        - mutualTLS: []
          operatorBearer: []
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A bounded fleet-health + circuit-breaker summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: |
        Mutual-TLS client-certificate authentication of the PEER PRODUCT (Sentinel
        or Delta). DECLARED here and APPLIED to every operation; certificate
        PROVISIONING and trust-store setup are DEFERRED to O-008 (F-034). Until then
        this scheme is a contract declaration, not a live enforced channel (honesty
        boundary a). mTLS establishes product identity only — it does not sign or
        authorize individual request bodies.

    hmacIngest:
      type: apiKey
      in: header
      name: X-Sentinel-Signature
      description: |
        Per-event HMAC-SHA256 body signature on the ingest seam, reusing the F-020
        webhook pattern (Anoryx-Sentinel/src/orchestration/webhooks/signer.py). The
        signer computes `HMAC-SHA256(signing_secret, "{timestamp}.{body}")` and sends
        the hex digest as `X-Sentinel-Signature: sha256=<hex>`, paired with the
        `X-Sentinel-Timestamp` header (Unix epoch seconds). Receivers MUST reject a
        timestamp outside a ±300s window (replay defense) and MUST recompute and
        compare the digest in constant time. NOTE: `apiKey` is the closest OpenAPI 3.1
        modelable type for an HMAC header; the real semantics (keyed signature over
        timestamp+body, not a static secret) are described here, not in the type. A
        present, well-formed signature is NOT a verified signature.

    serviceToken:
      type: http
      scheme: bearer
      description: |
        Delta service token (bearer) authorizing the distribution + query caller.
        Bearer-only: it authorizes the caller but does NOT sign request bodies (the
        embedded policy compact-JWS does that for distribution). Issued/rotated by
        deploy tooling; provisioning details are out of scope for O-001.
        NOTE (O-006): this COARSE shared `ORCH_SERVICE_TOKEN` gates BOTH the distribution
        GET-status (GET /v1/policies/distributions/{id}) and the distribution WRITE (POST
        /v1/policies/distributions) as a trusted multi-tenant relay — on the POST its inbound
        `tenant_id` is server-resolved from the signed body, not validated per-tenant (O-004
        LOW-1 and LOW-2 BOTH carried forward). For the NEW per-tenant read seams (/v1/events,
        /v1/bus/dlq) the coarse `ORCH_SERVICE_TOKEN` is legacy/not accepted: those seams
...
  parameters:
...
    DistributionId:
      name: distribution_id
      in: path
      required: true
      description: Opaque identifier returned by POST /v1/policies/distributions.
      schema:
        type: string
        format: uuid
        maxLength: 64
      example: "a7b8c9d0-1234-4ef0-9abc-def012345678"
...
  schemas:
...
    DistributionStatus:
      type: object
      additionalProperties: false
      required: [distribution_id, policy_id, state, created_at]
      description: |
        Distribution status for a submitted policy (read-only metadata). Aggregated
        `state` plus per-target states. No policy body is returned.
      properties:
        distribution_id:
          type: string
          format: uuid
          maxLength: 64
        policy_id:
          type: string
          format: uuid
          maxLength: 64
        state:
          type: string
          enum: [pending, distributed, partial, failed]
          description: Aggregate distribution state across all targets.
        created_at:
          type: string
          format: date-time
          maxLength: 64
        targets:
          type: array
          maxItems: 256
          items:
            $ref: '#/components/schemas/DistributionTargetStatus'
...